requirements.txt
numpy==1.21.2
pygame==2.1.2
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
